+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased][unreleased]
+### Added
+- Added `Fire\Post\Type::addSupport()`
+- Added `Fire\Post\Type::removeSupport()`
 
 ## [3.3.5] - 2021-04-15
 - Fixed `Fire\Post\page_id_for_type()` to get correct type even if no posts exist